// src/types/branch.ts
export interface Branch {
  id: number
  name: string
  branch_type: string
  street: string
  district: string
  city: string
  region: string
  country: string
  postal_code: string
  phone: string
  working_hours: string
  is_active: boolean
  location_link: string
  manager_name: string
  manager_phone: string
  has_fitting_room: boolean
  has_parking: boolean
  is_24_hours: boolean
  created_at: string
  updated_at: string
}